pyproject/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pandas
plotly
python-dotenv
pydantic
